requirements.txt
fastmcp
playwright
beautifulsoup4
lxml